reorganize benchmark to take capacity as input
The prior "input" scheme was a larger battery of random elements.
The interesting part of the LRU benchmark is not the throughput on
larger input sequences but the throughput on various sizes of the
fixed sized cache.

The results better exhibit the run time complexities of the
implementations and you can see how benoitc/lru gets slower
over a larger capacity (I'll touch on this in the analysis in some
child commit).