use a second ets table for ranks
This slows down `get/2` by a measurable but not terrible factor -
it takes away about 200K ips but that doesn't degrade poorly with
an increased capacity.

It's a huge boon for `put/2` though. Now `put/2` in the ets implementation
is about 2x faster than the gen_server example and is probably much
better behaved on parallel puts. This does beg some questions about
parallel puts or gets messing with the data integrity because these
operations are not atomic. This cache is not necessarily the highest
fidelity of data though so I suppose it doesn't really matter?